moving ipc deriving to trait
